Moved test helpers for `ImageCommentsViewController` into separate file
